+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 * Added `Document:keyInfo()`. It returns the last updated `DataStoreKeyInfo` returned from loading, saving, or closing the document. ([#50])
 * Migrations can now update the data mutably. Note: The value still needs to be returned from the migration function. ([#51])
 * Gracefully handle `validate` throwing an error in `Collection:load`. Before it would keep retrying the load and spam the console with confusing errors. ([#52])
+* Allow `defaultData` to be a function. The return value will be validated when a new document is created. ([#53])
 
 [#50]: https://github.com/nezuo/lapis/pull/50
 [#51]: https://github.com/nezuo/lapis/pull/51
 [#52]: https://github.com/nezuo/lapis/pull/52
+[#53]: https://github.com/nezuo/lapis/pull/53
 
 ### 0.3.0 - April 14, 2024
 * **BREAKING CHANGE**: `Collection:load` no longer caches promises. Each call will now return a unique promise and attempt to load the document separately. This is to fix an edge case that can result in data loss. More information can be found in the pull request. ([#48])